+++ src/dashboardWebView/components/SnippetsView/Item.tsx
@@ -1,7 +1,7 @@
 import { Messenger } from '@estruyf/vscode/dist/client';
 import { CodeIcon, DotsHorizontalIcon, PencilIcon, PlusIcon, TrashIcon } from '@heroicons/react/outline';
 import * as React from 'react';
-import { useCallback, useRef, useState } from 'react';
+import { useCallback, useMemo, useRef, useState } from 'react';
 import { useRecoilValue } from 'recoil';
 import { FeatureFlag } from '../../../components/features/FeatureFlag';
 import { FEATURE_FLAG } from '../../../constants';
@@ -31,8 +31,11 @@ export const Item: React.FunctionComponent<IItemProps> = ({ title, snippet }: Re
   const [ snippetTitle, setSnippetTitle ] = useState<string>('');
   const [ snippetDescription, setSnippetDescription ] = useState<string>('');
   const [ snippetOriginalBody, setSnippetOriginalBody ] = useState<string>('');
+  const [ mediaSnippet, setMediaSnippet ] = useState<boolean>(false);
 
   const formRef = useRef<SnippetFormHandle>(null);
+
+  const insertToContent = useMemo(() => viewData?.data?.filePath, [ viewData ]);
 
   const insertToArticle = () => {
     formRef.current?.onSave();
@@ -44,6 +47,7 @@ export const Item: React.FunctionComponent<IItemProps> = ({ title, snippet }: Re
     setSnippetTitle('');
     setSnippetDescription('');
     setSnippetOriginalBody('');
+    setMediaSnippet(false);
   };
 
   const onOpenEdit = useCallback(() => {
@@ -51,6 +55,7 @@ export const Item: React.FunctionComponent<IItemProps> = ({ title, snippet }: Re
     setSnippetDescription(snippet.description);
     setSnippetOriginalBody(typeof snippet.body === "string" ? snippet.body : snippet.body.join(`\n`));
     setShowEditDialog(true);
+    setMediaSnippet(!!snippet.isMediaSnippet);
   }, [snippet]);
   
   const onSnippetUpdate = useCallback(() => {
@@ -68,10 +73,15 @@ export const Item: React.FunctionComponent<IItemProps> = ({ title, snippet }: Re
 
     const snippetContents: Snippet = {
       ...crntSnippet,
-      fields,
       description: snippetDescription || '',
       body: snippetLines.length === 1 ? snippetLines[0] : snippetLines
     };
+
+    if (!mediaSnippet) {
+      snippetContents.fields = fields;
+    } else {
+      snippetContents.isMediaSnippet = true;
+    }
 
     // Check if new or update
     if (title === snippetTitle) {
@@ -84,7 +94,7 @@ export const Item: React.FunctionComponent<IItemProps> = ({ title, snippet }: Re
     Messenger.send(DashboardMessage.updateSnippet, { snippets });
 
     reset();
-  }, [settings?.snippets, title, snippetTitle, snippetDescription, snippetOriginalBody]);
+  }, [settings?.snippets, title, snippetTitle, snippetDescription, snippetOriginalBody, mediaSnippet]);
 
   const onDelete = useCallback(() => {
     const snippets = Object.assign({}, settings?.snippets || {});
@@ -108,7 +118,7 @@ export const Item: React.FunctionComponent<IItemProps> = ({ title, snippet }: Re
           features={mode?.features || []} 
           flag={FEATURE_FLAG.dashboard.snippets.manage}
           alternative={(
-            viewData?.data?.filePath ? (
+            insertToContent ? (
               <div className={`absolute top-4 right-4 flex flex-col space-y-4`}>
                 <div className="flex items-center border border-transparent group-hover:bg-gray-200 dark:group-hover:bg-vulcan-200 group-hover:border-gray-100 dark:group-hover:border-vulcan-50 rounded-full p-2 -mr-2 -mt-2">
                   <div className='group-hover:hidden'>
@@ -135,7 +145,7 @@ export const Item: React.FunctionComponent<IItemProps> = ({ title, snippet }: Re
 
               <div className='hidden group-hover:flex'>
                 {
-                  viewData?.data?.filePath && (
+                  insertToContent && !snippet.isMediaSnippet && (
                     <>
                       <QuickAction 
                         title={`Insert snippet`}
@@ -170,7 +180,7 @@ export const Item: React.FunctionComponent<IItemProps> = ({ title, snippet }: Re
           <FormDialog 
             title={`Insert snippet: ${title}`}
             description={`Insert the ${title.toLowerCase()} snippet into the current article`}
-            isSaveDisabled={!viewData?.data?.filePath}
+            isSaveDisabled={!insertToContent}
             trigger={insertToArticle}
             dismiss={() => setShowInsertDialog(false)}
             okBtnText='Insert'
@@ -200,6 +210,8 @@ export const Item: React.FunctionComponent<IItemProps> = ({ title, snippet }: Re
               title={snippetTitle}
               description={snippetDescription}
               body={snippetOriginalBody}
+              isMediaSnippet={mediaSnippet}
+              onMediaSnippetUpdate={(value: boolean) => setMediaSnippet(value)}
               onTitleUpdate={(value: string) => setSnippetTitle(value)}
               onDescriptionUpdate={(value: string) => setSnippetDescription(value)}
               onBodyUpdate={(value: string) => setSnippetOriginalBody(value)} />